+++ MeetingBar/StatusBarItemController.swift
@@ -1,5 +1,5 @@
 //
-//  StatusBarItemControler.swift
+//  StatusBarItemController.swift
 //  MeetingBar
 //
 //  Created by Andrii Leitsius on 12.06.2020.
@@ -15,10 +15,11 @@ import KeyboardShortcuts
 /**
  * creates the menu in the system status bar, creates the menu items and controls the whole lifecycle.
  */
-class StatusBarItemControler: NSObject, NSMenuDelegate {
+class StatusBarItemController: NSObject, NSMenuDelegate {
     var statusItem: NSStatusItem!
     var statusItemMenu: NSMenu!
     var menuIsOpen = false
+    var currentStatusBarEvent: EKEvent?
 
     var eventStore: EKEventStore!
     var calendars: [EKCalendar] = []
@@ -69,7 +70,6 @@ class StatusBarItemControler: NSObject, NSMenuDelegate {
     func statusMenuBarAction(sender _: NSStatusItem) {
         if !menuIsOpen, statusItem.menu == nil {
             let event = NSApp.currentEvent
-            NSLog("Event occured \(String(describing: event?.type.rawValue))")
 
             // Right button click
             if event?.type == NSEvent.EventType.rightMouseUp {
@@ -162,6 +162,13 @@ class StatusBarItemControler: NSObject, NSMenuDelegate {
                     button.image = NSImage(named: Defaults[.eventTitleIconFormat].rawValue)!
                 default:
                     button.image = NSImage(named: "iconCalendar")
+                }
+            }
+
+            if currentStatusBarEvent?.eventIdentifier != nextEvent?.eventIdentifier {
+                if nextEvent == nil || (nextEvent!).startDate.timeIntervalSinceNow <= 0 {
+                    currentStatusBarEvent = nextEvent
+                    updateMenu()
                 }
             }
 
@@ -324,6 +331,12 @@ class StatusBarItemControler: NSObject, NSMenuDelegate {
             toggleMeetingTitleVisibilityItem.setShortcut(for: .toggleMeetingTitleVisibilityShortcut)
             quickActionsItem.submenu!.addItem(toggleMeetingTitleVisibilityItem)
         }
+
+        let refreshSourcesItem = NSMenuItem()
+        refreshSourcesItem.title = "status_bar_section_refresh_sources".loco()
+        refreshSourcesItem.action = #selector(AppDelegate.refreshSources)
+        refreshSourcesItem.keyEquivalent = ""
+        quickActionsItem.submenu!.addItem(refreshSourcesItem)
     }
 
     func createBookmarksSection() {
@@ -401,7 +414,7 @@ class StatusBarItemControler: NSObject, NSMenuDelegate {
             image!.size = NSSize(width: 16, height: 16)
 
         // tested and verified
-        case .some(.meet):
+        case .some(.meet), .some(.meetStream):
             image = NSImage(named: "google_meet_icon")!
             image!.size = NSSize(width: 16, height: 13.2)
 
@@ -430,8 +443,9 @@ class StatusBarItemControler: NSObject, NSMenuDelegate {
             image = NSImage(named: "amazon_chime_icon")!
             image!.size = NSSize(width: 16, height: 16)
 
+        // tested and verified
         case .some(.ringcentral):
-            image = NSImage(named: "online_meeting_icon")!
+            image = NSImage(named: "ringcentral_icon")!
             image!.size = NSSize(width: 16, height: 16)
 
         // tested and verified
@@ -444,8 +458,9 @@ class StatusBarItemControler: NSObject, NSMenuDelegate {
             image = NSImage(named: "gotowebinar_icon")!
             image!.size = NSSize(width: 16, height: 16)
 
+        // tested and verified
         case .some(.bluejeans):
-            image = NSImage(named: "online_meeting_icon")!
+            image = NSImage(named: "bluejeans_icon")!
             image!.size = NSSize(width: 16, height: 16)
 
         // tested and verified
@@ -534,16 +549,42 @@ class StatusBarItemControler: NSObject, NSMenuDelegate {
             image!.size = NSSize(width: 16, height: 16)
 
         // tested and verified
+        case .some(.zhumu):
+            image = NSImage(named: "zhumu_icon")!
+            image!.size = NSSize(width: 16, height: 16)
+
+        // tested and verified
+        case .some(.lark):
+            image = NSImage(named: "lark_icon")!
+            image!.size = NSSize(width: 16, height: 16)
+
+        // tested and verified
+        case .some(.feishu):
+            image = NSImage(named: "feishu_icon")!
+            image!.size = NSSize(width: 16, height: 16)
+
+        // tested and verified
+        case .some(.vimeo_showcases):
+            image = NSImage(named: "vimeo_icon")!
+            image!.size = NSSize(width: 16, height: 16)
+
+        // tested and verified
+        case .some(.ovice):
+            image = NSImage(named: "ovice_icon")!
+            image!.size = NSSize(width: 16, height: 16)
+
+        case .some(.facetime):
+            image = NSImage(named: "facetime_icon")!
+            image!.size = NSSize(width: 16, height: 16)
+
+        // tested and verified
         case .none:
             image = NSImage(named: "no_online_session")!
             image!.size = NSSize(width: 16, height: 16)
 
+        // tested and verified
         case .some(.vonageMeetings):
-            image = NSImage(named: "online_meeting_icon")!
-            image!.size = NSSize(width: 16, height: 16)
-
-        case .some(.meetStream):
-            image = NSImage(named: "online_meeting_icon")!
+            image = NSImage(named: "vonage_icon")!
             image!.size = NSSize(width: 16, height: 16)
 
         case .some(.url):
@@ -597,7 +638,7 @@ class StatusBarItemControler: NSObject, NSMenuDelegate {
 
         switch Defaults[.timeFormat] {
         case .am_pm:
-            eventTimeFormatter.dateFormat = "hh:mm a"
+            eventTimeFormatter.dateFormat = "h:mm a  "
         case .military:
             eventTimeFormatter.dateFormat = "HH:mm"
         }
@@ -915,7 +956,9 @@ class StatusBarItemControler: NSObject, NSMenuDelegate {
     }
 
     func createPreferencesSection() {
-        if removePatchVerion(Defaults[.appVersion]) > removePatchVerion(Defaults[.lastRevisedVersionInChangelog]) {
+        let showChangelogItem = compareVersions(Defaults[.appVersion], Defaults[.lastRevisedVersionInChangelog])
+
+        if showChangelogItem {
             let changelogItem = statusItemMenu.addItem(
                 withTitle: "status_bar_whats_new".loco(),
                 action: #selector(AppDelegate.openChangelogWindow),